+++ lua/smark/init.lua
@@ -10,18 +10,25 @@ function M.setup(opts)
 		--Keymapping settings for list action commands.
 		--Set to false to disable.
 		mappings = {
-			--Format the current list block to be clean / correct
+			--Format the current list block to be clean / correct.
 			format_list = "<leader>lf",
-			--Switch between ordered / unordered list types
+			--Switch between ordered / unordered list types.
 			toggle_ordered = "<leader>lo",
-			--Toggle the completion status of a task list item
+			--Toggle the completion status of a task list item.
 			toggle_completion = "<leader>lx",
-			--Toggle between plain and task list items
+			--Toggle between plain and task list items.
 			toggle_task = "<leader>lt",
 		},
+
+		--Following the starting line of a list item, only contiguous lines that
+		--start with at least one whitespace character can be considered as part of
+		--a multi-line list item.
+		multiline_requires_whitespace = false,
 	}
 
 	opts = vim.tbl_deep_extend("force", default_opts, opts or {})
+
+	buffer.multiline_requires_whitespace = opts.multiline_requires_whitespace
 
 	vim.api.nvim_create_autocmd("FileType", {
 		pattern = { "markdown", "text" },
@@ -70,7 +77,7 @@ function M.setup(opts)
 end
 
 function callback.insert_newline()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -88,14 +95,14 @@ function callback.insert_newline()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		new_line_at_cursor
 	)
 	vim.api.nvim_win_set_cursor(0, { cursor_coords.row, cursor_coords.col })
 end
 
 function callback.insert_indent()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -113,14 +120,14 @@ function callback.insert_indent()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 	vim.api.nvim_win_set_cursor(0, { cursor_coords.row, cursor_coords.col })
 end
 
 function callback.insert_unindent()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -143,14 +150,14 @@ function callback.insert_unindent()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 	vim.api.nvim_win_set_cursor(0, { cursor_coords.row, cursor_coords.col })
 end
 
 function callback.normal_indent()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -171,13 +178,13 @@ function callback.normal_indent()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 end
 
 function callback.normal_unindent()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -198,7 +205,7 @@ function callback.normal_unindent()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 end
@@ -226,7 +233,7 @@ function callback.normal_unindent_op()
 end
 
 function callback.normal_o()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -243,7 +250,7 @@ function callback.normal_o()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		true
 	)
 	vim.api.nvim_win_set_cursor(0, { cursor_coords.row, cursor_coords.col })
@@ -251,7 +258,7 @@ function callback.normal_o()
 end
 
 function callback.normal_format()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -266,13 +273,13 @@ function callback.normal_format()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 end
 
 function callback.normal_ordered()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -288,13 +295,13 @@ function callback.normal_ordered()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 end
 
 function callback.normal_completion()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -310,13 +317,13 @@ function callback.normal_completion()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 end
 
 function callback.normal_task()
-	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start =
+	local li_block_bounds, li_block, read_time_lines, li_cursor_coords, to_put_separator_at_start, to_put_separator_at_end =
 		buffer.get_list_block_around_cursor()
 
 	if li_block_bounds == nil then
@@ -332,7 +339,7 @@ function callback.normal_task()
 		li_block_bounds,
 		cursor_coords,
 		to_put_separator_at_start,
-		false,
+		to_put_separator_at_end,
 		false
 	)
 end